+++ lib/src/brotli/ffi/library.dart
@@ -16,11 +16,29 @@ import 'types.dart';
 class BrotliLibrary
     with OpenLibrary, BrotliConstants, BrotliFunctions, BrotliTypes {
   /// Library path the user can define to override normal resolution.
-  static String userDefinedLibraryPath;
+  static String _userDefinedLibraryPath;
+
+  /// Return the library path defined by the user.
+  static String get userDefinedLibraryPath => _userDefinedLibraryPath;
+
+  /// Set the library [path] defined by the user.
+  ///
+  /// Throw a [StateError] if this library has already been initialized.
+  static set userDefinedLibraryPath(String path) {
+    if (_initialized == true) {
+      throw StateError('BrotliLibrary already initialized.');
+    }
+    _userDefinedLibraryPath = path;
+  }
 
   /// Singleton instance.
   static final BrotliLibrary _instance =
-      BrotliLibrary._(userDefinedLibraryPath);
+      BrotliLibrary._(_userDefinedLibraryPath);
+
+  /// Tracks library init state.
+  ///
+  /// Set to [:true:] if this library is opened and all functions are resolved.
+  static bool _initialized = false;
 
   /// Dart native library object.
   DynamicLibrary _libraryImpl;
@@ -39,5 +57,6 @@ class BrotliLibrary
   BrotliLibrary._(String libraryPath) {
     _libraryImpl = openLibrary(path: libraryPath);
     resolveFunctions(_libraryImpl);
+    _initialized = true;
   }
 }