+++ lib/src/brotli/stubs/codec.dart
@@ -6,6 +6,18 @@ import 'dart:convert';
 
 import '../codec.dart';
 
+/// Library function used by [BrotliCodec] static function that gets the custom
+/// library path.
+///
+/// This is the stubbed version that be a no-op.
+String brotliGetLibraryPath() => null;
+
+/// Library function used by [BrotliCodec] static function that sets a custom
+/// library path.
+///
+/// This is the stubbed version that be a no-op.
+void brotliSetLibraryPath(String path) => null;
+
 /// Extension that provides suitable stubs for [BrotliCodec]s in non-IO
 /// environments.
 ///