+++ lib/src/zstd/codec.dart
@@ -13,6 +13,14 @@ import 'version.dart';
 /// The [ZstdCodec] encodes/decodes raw bytes using the Zstd (ZStandard)
 /// algorithm.
 class ZstdCodec extends Codec<List<int>, List<int>> {
+  /// Return the library path [String] or [:null:] if not set.
+  static String get libraryPath => zstdGetLibraryPath();
+
+  /// Set the custom library [path]
+  ///
+  /// Throw a [StateError] if the library has already been initialized.
+  static set libraryPath(String path) => zstdSetLibraryPath(path);
+
   /// The compression-[level] can be set in the range of `0..16`, with
   /// 0 (fast mode) being the default compression level.
   final int level;
@@ -45,10 +53,8 @@ class ZstdCodec extends Codec<List<int>, List<int>> {
   ZstdCodec(
       {this.level = ZstdOption.defaultLevel,
       this.inputBufferLength = CodecBufferHolder.autoLength,
-      this.outputBufferLength = CodecBufferHolder.autoLength,
-      String libraryPath}) {
+      this.outputBufferLength = CodecBufferHolder.autoLength}) {
     validateZstdLevel(level);
-    if (libraryPath != null) userDefinedLibraryPath = libraryPath;
   }
 
   /// Internal Constructor for building the [zstd] instance.